requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
Flask-JWT-Extended==4.3.1
celery==4.4.2
redis==3.5.3
ffmpeg-python==0.2.0